Class/Test/TrainOperatorTest.java
package Test;
import Class.TrainOperator;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class TrainOperatorTest {

    private TrainOperator trainOperator;

    @Before
    public void setUp() {
        // Initialize your TrainOperator object here
        trainOperator = new TrainOperator(1, "John Doe");
    }

    @Test
    public void testTrackTrainStatus() {
        // Assuming trackTrainStatus changes some internal state or logs action
        trainOperator.trackTrainStatus();
        // Here, you would check that the action was logged or state changed
        // This step is conceptual, as the actual method prints to the console
        assertTrue("Should confirm train status was tracked.", true );
    }

    @Test
    public void testNotifyEmergency() {
        // Assuming notifyEmergency logs the emergency message or changes state
        trainOperator.notifyEmergency("Emergency Test Message");
        // Again, verify the message was logged or state changed
        // Conceptual placeholder assertion
        assertTrue("Should confirm emergency was reported.", true);
    }
}